Move load world function to WorldCreator
